Add index.html
Which is a redirection page which automatically redirects to either about
or modes, depending on if this is the user's first visit.... But since I
don't want to start writing scripts right now (who wants to deal with
cookies anyway??), for now it assumes all users are new and redirects to
about.

Redirection is performed in three ways. First, a meta-refresh sends the
browser to about after zero seconds. Second, a JS event sets the window
location to about once the body loads. Finally, if neither of those work,
a very barebones page appears, prompting the user to click a link to
proceed to about manually.

Additionally, there is a canonical page link tag, as a fourth suggestion
for automated crawlers.