Choose start page countdown date dynamically
The rule is as follows: All countdowns go to 8:30AM, with zero seconds and
zero milliseconds. If the current time is less than the current time with
the hour set to 8AM, the minute set to 30, and the second and millisecond
set to 0, then count down to 8:30. Else, add one day to the target, and
count down to 8:30 tomorrow.