Changes from Alex: Improve the modal displays
The height is increased to add more space, which is used to make the
checkboxes cartoonishly large (Hey, I like cartoons), and to move the time
slider output display underneath the slider. The slider's width adjustment
isn't changed, so it doesn't fill the page width anymore, but it still
looks good and works -- I can fix it later if need be.

Other changes include a bugfix to bring clicking out of modal boxes to all
of the boxes, and a transparent background for all of them as well.