Improve style of new row button
This was the only egregious style error that was holding me back from
pushing out an update.

I'm thinking that we might want to invert the colors on the controls (like
I did for advice), and that I'd like to restyle the time control for sure,
but for now this is good.