Fix negative time since intake in caffeine table - Fixes #8
Entries which would previously have been used as being in the future are
now set so they occur on a previous day.

This shouldn't be an issue, but it is possible that times are set several
days in the future. If this is a problem, use a while loop instead of ifs
to check for future intake.