Further changes from Meidi: Improve summary page
Also add a welcome page. This welcome page is merely a manual redirect,
and now receives the automatic redirect from the index.

Improvements to the summary page include the expansion of the summary to
cover the custom mode.

Meidi says the following regarding these changes:

Meidi Peng: version with welcome page and improvements on summary page
Meidi Peng: hope this will make the whole project a little bit easier to understand...

Signed-off-by: Ryan Hodin <ryan951357@gmail.com>