Add handling for *teen postfixes - Fixes #6
This also correct the handling of "11" - It is "eleventh", not "elevenst".

The fix implemented relies on all numbers that have a 1 in the second to
last position properly having an ordinal postfix "th" - tenth, eleventh,
twelfth, thirteenth, fourteenth, fifteenth, sixteenth, seventeenth,
eighteenth, nineteenth. It does not, however, account for improperly
formed strings being passed - "1." would become "1.th", which is odd to
say the least.

This brings me to my other point. I have decided that it is part of the
contract of using the postfixer that one is only passing numbers, or
objects whose string representation ends with a number. Therefore, while
our error handling is nice, it need not be robust - We make no guarantees
regarding the output of the function when an improper string is passed, so
while its nice of us to return the empty string, we are allowed to return
"th" - Thus letting us not examine the last letter in the "1x" check. This
is done knowing that it is a function to be used over a small codebase, so
knowing our inputs is easy, and tracking down a bug caused by this
behavior is also easy.

If you are looking at this commit message because this somehow became the
basis for a much larger codebase and you saw this deficit in error
handling and ran 'blame' to find out just who caused you all this
heartache, I sincerely apologize to you.

Yeesh, this was a long commit message....

Signed-off-by: Ryan Hodin <ryan951357@gmail.com>